fix(ci): respect layer boundary for URL normalization
